Major changes to handling login API responses. Renamed main Launcher folder to ToontownLauncher.
