+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = "Launcher"
+rootProject.name = "ToontownLauncher"
