--- src/main/java/com/example/Launcher/patterns/command/Command.java
@@ -1,6 +1,0 @@
-package com.example.Launcher.patterns.command;
-
-public interface Command {
-    void execute();
-    // void undo();  // Optional, to support undo functionality
-}
