--- src/main/java/com/example/Launcher/patterns/command/RemoveToonCommand.java
@@ -1,5 +1,0 @@
-package com.example.Launcher.patterns.command;
-
-public class RemoveToonCommand {
-    // needs to be implemented
-}
