--- src/main/java/com/example/Launcher/patterns/strategy/PlayStrategy.java
@@ -1,7 +1,0 @@
-package com.example.Launcher.patterns.strategy;
-
-import com.example.Launcher.models.Toon;
-
-public interface PlayStrategy {
-    void play(Toon toon);
-}
