--- src/main/java/com/example/Launcher/utils/ui/AlertManager.java
@@ -1,14 +1,0 @@
-package com.example.Launcher.utils.ui;
-
-import javafx.scene.control.Alert;
-
-public class AlertManager {
-
-    public static void showAlert(Alert.AlertType alertType, String title, String content) {
-        Alert alert = new Alert(alertType);
-        alert.setTitle(title);
-        alert.setHeaderText(null);
-        alert.setContentText(content);
-        alert.showAndWait();
-    }
-}
