--- src/main/java/com/example/Launcher/utils/ui/StyleManager.java
@@ -1,10 +1,0 @@
-package com.example.Launcher.utils.ui;
-
-import javafx.scene.Scene;
-
-public class StyleManager {
-
-    public void applyStylesheet(Scene scene, String stylesheetPath) {
-        scene.getStylesheets().add(getClass().getResource(stylesheetPath).toExternalForm());
-    }
-}
