+++ src/main/java/com/example/ToontownLauncher/patterns/command/Command.java
@@ -1,0 +1,6 @@
+package com.example.ToontownLauncher.patterns.command;
+
+public interface Command {
+    void execute();
+    // void undo();  // Optional, to support undo functionality
+}
