+++ src/main/java/com/example/ToontownLauncher/patterns/command/RemoveToonCommand.java
@@ -1,0 +1,5 @@
+package com.example.ToontownLauncher.patterns.command;
+
+public class RemoveToonCommand {
+    // needs to be implemented
+}
