--- src/main/resources/com/example/Launcher/Display.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.layout.HBox?>
-
-<VBox xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="com.example.Launcher.controllers.DisplayController"
-    alignment="CENTER" spacing="10" prefHeight="400" prefWidth="400">
-
-    <ListView fx:id="toonsListView" />
-
-    <HBox alignment="CENTER" spacing="10">
-        <Button text="Add Toon" onAction="#addToonClicked" />
-        <Button fx:id="playButton" text="Play" onAction="#playClicked" />
-    </HBox>
-
-</VBox>
