--- src/main/resources/com/example/Launcher/Form.fxml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.layout.HBox?>
-
-<VBox xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="com.example.Launcher.controllers.eventhandlers.FormEventHandlers"
-      alignment="CENTER" spacing="10" prefHeight="200" prefWidth="250">
-
-   <Label text="Add Toon" />
-
-   <TextField fx:id="nameField" promptText="Name" />
-   <TextField fx:id="usernameField" promptText="Username" />
-   <PasswordField fx:id="passwordField" promptText="Password" />
-
-   <HBox spacing="10" alignment="CENTER">
-      <Button text="Save" onAction="#saveToon" />
-      <Button text="Cancel" onAction="#closeForm" />
-   </HBox>
-
-</VBox>
